requirements.txt
numpy
opencv-python
matplotlib
numba
scikit-learn
Pillow

